requirements.txt
# Core dependencies
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0

# Data processing and file formats
pyarrow>=7.0.0
openpyxl>=3.0.9

# Chinese text processing
pypinyin==0.53.0
char_converter==0.40
nltk==3.9.1

# Machine learning
scikit-learn==1.5.2
xgboost==2.1.3

# Network analysis and clustering
networkx==3.4.2
leidenalg==0.10.2
python-igraph==0.11.0

# Word embeddings
gensim==4.4.3

# Utilities
tqdm==4.67.1

# Visualization
matplotlib==3.10.0